+++ docs/start/flowgrpo_quickstart.md
@@ -99,7 +99,7 @@ WORKSPACE              # base directory for data (default: $HOME)
 
 ## Step 3: Perform FlowGRPO training
 
-The provided example script launches `python3 -m verl_omni.trainer.main_flowgrpo` with the FlowGRPO-specific config needed for this OCR task:
+The provided example script launches `python3 -m verl_omni.trainer.diffusion.main_flowgrpo` with the FlowGRPO-specific config needed for this OCR task:
 
 - `algorithm.adv_estimator=flow_grpo`
 - `actor_rollout_ref.rollout.name=vllm_omni`
