+++ examples/flowgrpo_trainer/run_qwen_image_ocr_lora_async_reward.sh
@@ -20,7 +20,7 @@ ENGINE=vllm_omni
 REWARD_ENGINE=vllm
 
 
-python3 -m verl_omni.trainer.main_flowgrpo \
+python3 -m verl_omni.trainer.diffusion.main_flowgrpo \
     algorithm.adv_estimator=flow_grpo \
     data.train_files=$ocr_train_path \
     data.val_files=$ocr_test_path \
