+++ tests/agent_loop/test_diffusion_agent_loop.py
