+++ tests/reward_loop/test_visual_reward_manager.py
