+++ tests/special_e2e/run_flowgrpo_trainer_diffusers.sh
@@ -32,7 +32,7 @@ micro_bsz=$((micro_bsz_per_gpu * NUM_GPUS))
 mini_bsz=${micro_bsz}
 train_batch_size=$((mini_bsz * n_resp_per_prompt))
 
-python3 -m verl_omni.trainer.main_flowgrpo \
+python3 -m verl_omni.trainer.diffusion.main_flowgrpo \
     algorithm.adv_estimator=flow_grpo \
     data.train_files=${dummy_train_path} \
     data.val_files=${dummy_test_path} \
