+++ tests/special_sanity/check_docstrings.py
@@ -89,10 +89,10 @@ def main():
     """Main function to check docstrings in specified files."""
 
     files_to_check = [
-        "verl_omni/trainer/main_flowgrpo.py",
+        "verl_omni/trainer/diffusion/main_flowgrpo.py",
         "verl_omni/trainer/diffusion/ray_diffusion_trainer.py",
         "verl_omni/trainer/diffusion/diffusion_algos.py",
-        "verl_omni/experimental/agent_loop/diffusion_agent_loop.py",
+        "verl_omni/agent_loop/diffusion_agent_loop.py",
     ]
 
     script_dir = os.path.dirname(os.path.abspath(__file__))
