+++ verl_omni/_patch.py
@@ -28,8 +28,8 @@ def _patch_diffusion_agent_loop() -> None:
     (e.g. inside ``AgentLoopManager.__init__``) resolve to them.
     TODO (mike): to be dropped
     """
-    import verl_omni.experimental.agent_loop.diffusion_agent_loop as _omni_dal
-    import verl_omni.experimental.agent_loop.single_turn_agent_loop  # noqa: F401
+    import verl_omni.agent_loop.diffusion_agent_loop as _omni_dal
+    import verl_omni.agent_loop.single_turn_agent_loop  # noqa: F401
 
     sys.modules["verl.experimental.agent_loop.diffusion_agent_loop"] = _omni_dal
 
@@ -62,15 +62,15 @@ def _patch_vllm_omni_replica() -> None:
 
 def _patch_diffusers_model() -> None:
     """Alias ``verl.models.diffusers_model`` (and its ``base`` / ``utils``
-    submodules) to verl-omni's ``verl_omni.custom_pipelines`` so
+    submodules) to verl-omni's ``verl_omni.pipelines`` so
     that upstream code using the verl-side ``DiffusionModelBase`` registry
     (e.g. ``verl/workers/engine/fsdp/diffusers_impl.py``) resolves to the
     verl-omni registry where pipelines are actually registered.
     TODO (mike): to be dropped
     """
-    import verl_omni.custom_pipelines as _omni_pkg
-    import verl_omni.custom_pipelines.base as _omni_base
-    import verl_omni.custom_pipelines.utils as _omni_utils
+    import verl_omni.pipelines as _omni_pkg
+    import verl_omni.pipelines.model_base as _omni_base
+    import verl_omni.pipelines.utils as _omni_utils
 
     sys.modules["verl.models.diffusers_model"] = _omni_pkg
     sys.modules["verl.models.diffusers_model.base"] = _omni_base
@@ -116,7 +116,7 @@ def _patch_visual_reward_manager() -> None:
     import verl.experimental.reward_loop.reward_manager  # noqa: F401 — triggers verl's @register("visual")
     from verl.experimental.reward_loop.reward_manager.registry import REWARD_MANAGER
 
-    from verl_omni.experimental.reward_loop.reward_manager.visual import VisualRewardManager as _OmniVisual
+    from verl_omni.reward_loop.reward_manager.visual import VisualRewardManager as _OmniVisual
 
     REWARD_MANAGER["visual"] = _OmniVisual
