+++ verl_omni/agent_loop/single_turn_agent_loop.py
@@ -19,7 +19,7 @@ from uuid import uuid4
 from verl.experimental.agent_loop.agent_loop import AgentLoopBase, register
 from verl.utils.profiler import simple_timer
 
-from verl_omni.experimental.agent_loop.diffusion_agent_loop import DiffusionAgentLoopOutput
+from verl_omni.agent_loop.diffusion_agent_loop import DiffusionAgentLoopOutput
 
 logger = logging.getLogger(__file__)
 logger.setLevel(os.getenv("VERL_LOGGING_LEVEL", "WARN"))
