+++ verl_omni/pipelines/__init__.py
@@ -12,6 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .qwen_image import *  # noqa: F401, F403
+from .qwen_image_flow_grpo import *  # noqa: F401, F403
 
-__all__ = qwen_image.__all__
+__all__ = qwen_image_flow_grpo.__all__
