+++ verl_omni/reward_loop/reward_manager/visual.py
